+++ packages/backend/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './logger';
 export * from './ajv';
 export * from './normalizeUrl';
+export * from './env';