+++ packages/frontend/pages/_document.page.tsx
@@ -1,4 +1,5 @@
 import NextDocument, { Html, Head, Main, NextScript } from 'next/document';
+import { ColorModeScript } from '@chakra-ui/react';
 
 export default class Document extends NextDocument {
   render() {
@@ -119,6 +120,7 @@ export default class Document extends NextDocument {
           />
         </Head>
         <body>
+          <ColorModeScript type="cookie" />
           <Main />
           <NextScript />
         </body>