+++ packages/frontend/utils/ClientRender.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import { ReactFCC } from '@utils/react';
 import { useIsClient } from 'usehooks-ts';
 
-export const ClientRender: React.FC = ({ children }) => {
+export const ClientRender: ReactFCC = ({ children }) => {
   const isClient = useIsClient();
   return isClient ? (children as JSX.Element) : null;
 };