+++ packages/frontend/utils/react.ts
@@ -1,0 +1,3 @@
+import React, { PropsWithChildren } from 'react';
+
+export type ReactFCC<T = {}> = React.FC<PropsWithChildren<T>>;