+++ CHANGELOG.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-05-22
+
+### Added
+
+-   Auto-expand parent collection when adding children to it
+-   Support `DB_PASS_FILE` and `COOKIE_SECRET_FILE` env variables
+
+### Changed
+
+-   Migrated to React 18 and Chakra UI v2
+-   Bumped all dependencies
+
+### Fixed
+
+-   About page: show short commit SHA
+-   Password page: disable form in demo mode
+-   Main page: fixed column width jumping on initial render
+-   Settings page: fix empty SSR render
+
 ## [0.1.1] - 2022-05-14
 
 ### Fixed
@@ -19,7 +38,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   The entire app!
 
-[Unreleased]: https://github.com/frysztak/orpington-news/compare/0.1.1...HEAD
+[Unreleased]: https://github.com/frysztak/orpington-news/compare/0.2.0...HEAD
+
+[0.2.0]: https://github.com/frysztak/orpington-news/compare/0.1.1...0.2.0
 
 [0.1.1]: https://github.com/frysztak/orpington-news/compare/0.1.0...0.1.1